# web/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Koritsu API",
  "description": "Two entrances, one set of handlers: `/api` for the site (cookie session) and `/api/v1` for external clients (bearer token). Only `/api/v1` is versioned and documented.",
  "version": "2.0.0a5.4"
 },
 "paths": {
...
  "/api/projects": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Create a project in a workspace",
    "description": "Creates a project and its directory on the volume. Multipart: the name and, optionally, either a DOCX template or the id of one of your saved templates (`template_id`); without either the report is built from scratch. `module` says which module the work is done with (from GET /api/modules) and may be left out. Editor role in the workspace. 400 bad_template, 400 unknown_module, 403 forbidden, 404 not_found, 409 project_exists.",
    "operationId": "create_project",
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_create_project"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Project"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "List projects of a workspace",
    "description": "Lists the projects of one workspace; `trash=true` lists the ones in the trash instead, `module` narrows the list down to the works done with one module. 400 invalid_id, 400 unknown_module, 404 not_found.",
    "operationId": "list_projects",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Workspace Id"
      }
     },
     {
      "name": "trash",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Trash"
      }
     },
     {
      "name": "module",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Module"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Projects"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}": {
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "One project",
    "description": "One project: name, workspace, the tag keys of its manifest and the size of its directory. Paths are never returned. 400 invalid_id, 404 not_found, 409 in_trash.",
    "operationId": "get_project",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Project"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "projects"
    ],
    "summary": "Rename a project or say which module it is done with",
    "description": "Changes the name of a project, in the database and in its settings on the volume, and the module it is done with. Both fields are optional; an empty body changes nothing. Editor role. 400 unknown_module, 403 forbidden, 404 not_found, 409 in_trash, 422 validation_failed.",
    "operationId": "rename_project",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectPatchIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Rename Project"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "projects"
    ],
    "summary": "Move a project to the trash",
    "description": "Moves a project to the trash; its directory stays on the volume until purge_after. Editor role. 403 forbidden, 404 not_found, 409 in_trash.",
    "operationId": "trash_project",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Trash Project"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/projects/{project_id}/runs": {
   "post": {
    "tags": [
     "projects"
    ],
    "summary": "Write down a module run of this project",
    "description": "Adds one entry to the run journal of a project: which module was started, what it is called and what it produced. The answer carries `n`, the sequence number of this run of this module in this project, which the interface uses to name a run that was left unnamed. Editor role. 400 invalid_id, 400 unknown_module, 403 forbidden, 404 not_found.",
    "operationId": "create_project_run",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectRunIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectRunOut"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "projects"
    ],
    "summary": "Run journal of a project",
    "description": "Everything that has been started in this project: which module, when, under what name, and what it produced. `module` narrows the list down to one module; `sort` orders it: newest first by default, then oldest, by name or by module. Viewer role. 400 invalid_id, 400 unknown_module, 404 not_found.",
    "operationId": "list_project_runs",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     },
     {
      "name": "module",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Module"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "new",
        "old",
        "name",
        "module"
       ],
       "type": "string",
       "default": "new",
       "title": "Sort"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ProjectRunOut"
         },
         "title": "Response List Project Runs"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/projects/{project_id}/runs/{run_id}": {
   "delete": {
    "tags": [
     "projects"
    ],
    "summary": "Remove one entry from the run journal",
    "description": "Removes a run from the journal of the project: this is how a diagram is deleted from a work. What the run produced stays on the volume: an artifact is addressed by its content and may be the value of a tag, so deleting it here would knock a picture out of a finished document. Editor role. 400 invalid_id, 403 forbidden, 404 not_found.",
    "operationId": "delete_project_run",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/modules": {
   "get": {
    "tags": [
     "modules"
    ],
    "summary": "Modules the interface may show",
    "description": "Every module the interface is allowed to show, in sidebar order: id, English title and where its routes live. Modules that are not ready are absent from the list entirely; there is no placeholder for them.",
    "operationId": "list_modules",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "additionalProperties": true,
          "type": "object"
         },
         "type": "array",
         "title": "Response List Modules"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects": {
   "post": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "Create a project in a workspace",
    "description": "Creates a project and its directory on the volume. Multipart: the name and, optionally, either a DOCX template or the id of one of your saved templates (`template_id`); without either the report is built from scratch. `module` says which module the work is done with (from GET /api/modules) and may be left out. Editor role in the workspace. 400 bad_template, 400 unknown_module, 403 forbidden, 404 not_found, 409 project_exists.",
    "operationId": "create_project_v1",
    "requestBody": {
     "required": true,
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_create_project_v1"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Create Project V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "List projects of a workspace",
    "description": "Lists the projects of one workspace; `trash=true` lists the ones in the trash instead, `module` narrows the list down to the works done with one module. 400 invalid_id, 400 unknown_module, 404 not_found.",
    "operationId": "list_projects_v1",
    "parameters": [
     {
      "name": "workspace_id",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Workspace Id"
      }
     },
     {
      "name": "trash",
      "in": "query",
      "required": false,
      "schema": {
       "type": "boolean",
       "default": false,
       "title": "Trash"
      }
     },
     {
      "name": "module",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Module"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Projects V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}": {
   "get": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "One project",
    "description": "One project: name, workspace, the tag keys of its manifest and the size of its directory. Paths are never returned. 400 invalid_id, 404 not_found, 409 in_trash.",
    "operationId": "get_project_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Get Project V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "Rename a project or say which module it is done with",
    "description": "Changes the name of a project, in the database and in its settings on the volume, and the module it is done with. Both fields are optional; an empty body changes nothing. Editor role. 400 unknown_module, 403 forbidden, 404 not_found, 409 in_trash, 422 validation_failed.",
    "operationId": "rename_project_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectPatchIn"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Rename Project V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "delete": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "Move a project to the trash",
    "description": "Moves a project to the trash; its directory stays on the volume until purge_after. Editor role. 403 forbidden, 404 not_found, 409 in_trash.",
    "operationId": "trash_project_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Trash Project V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
  "/api/v1/projects/{project_id}/runs": {
   "post": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "Write down a module run of this project",
    "description": "Adds one entry to the run journal of a project: which module was started, what it is called and what it produced. The answer carries `n`, the sequence number of this run of this module in this project, which the interface uses to name a run that was left unnamed. Editor role. 400 invalid_id, 400 unknown_module, 403 forbidden, 404 not_found.",
    "operationId": "create_project_run_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProjectRunIn"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ProjectRunOut"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "Run journal of a project",
    "description": "Everything that has been started in this project: which module, when, under what name, and what it produced. `module` narrows the list down to one module; `sort` orders it: newest first by default, then oldest, by name or by module. Viewer role. 400 invalid_id, 400 unknown_module, 404 not_found.",
    "operationId": "list_project_runs_v1",
    "parameters": [
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     },
     {
      "name": "module",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "default": "",
       "title": "Module"
      }
     },
     {
      "name": "sort",
      "in": "query",
      "required": false,
      "schema": {
       "enum": [
        "new",
        "old",
        "name",
        "module"
       ],
       "type": "string",
       "default": "new",
       "title": "Sort"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/ProjectRunOut"
         },
         "title": "Response List Project Runs V1"
        }
       }
      }
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/projects/{project_id}/runs/{run_id}": {
   "delete": {
    "tags": [
     "v1",
     "projects"
    ],
    "summary": "Remove one entry from the run journal",
    "description": "Removes a run from the journal of the project: this is how a diagram is deleted from a work. What the run produced stays on the volume: an artifact is addressed by its content and may be the value of a tag, so deleting it here would knock a picture out of a finished document. Editor role. 400 invalid_id, 403 forbidden, 404 not_found.",
    "operationId": "delete_project_run_v1",
    "parameters": [
     {
      "name": "run_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Run Id"
      }
     },
     {
      "name": "project_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Project Id"
      }
     }
    ],
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "default": {
      "description": "Any refusal: one shape, machine-readable code",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ErrorOut"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "Body_create_project": {
    "properties": {
     "workspace_id": {
      "type": "string",
      "title": "Workspace Id"
     },
     "name": {
      "type": "string",
      "title": "Name",
      "default": ""
     },
     "module": {
      "type": "string",
      "title": "Module",
      "default": ""
     },
     "template": {
      "anyOf": [
       {
        "type": "string",
        "contentMediaType": "application/octet-stream"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template"
     },
     "template_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Template Id"
     }
    },
    "type": "object",
    "required": [
     "workspace_id"
    ],
    "title": "Body_create_project"
   },
...
   "ErrorOut": {
    "properties": {
     "error": {
      "$ref": "#/components/schemas/ErrorBody"
     }
    },
    "type": "object",
    "required": [
     "error"
    ],
    "title": "ErrorOut",
    "description": "Единственная форма отказа службы. Другой нет ни у одного маршрута."
   },
...
  }
 },
...
}